Added "Workout" node to enclose entire workout.
I added the Workout node because SetList can occur within parts of the
workout and didn't seem appropriate as the main parent node. The other files
in the checkin are the results of changes that needed to be made from adding
the Workout node to the AST through swim_parser.